Change font for org-date to have more contrast (#45)
* Change font for org-date to have more contrast

The current value for org-date is darktooth-faded_blue which has
very little contrast. My expectation is that dates stand out
because they are important. All the other variants of blue can not
be used (either to dark or conflict with level 3), that's why I
suggest an aqua variant here: darktooth-neutral_aqua.

* Changing foreground of org-date to faded_aqua

Result of discussion in pull request.